+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/rbel/RbelMessageValidator.java
@@ -6,13 +6,12 @@ package de.gematik.test.tiger.lib.rbel;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
-import com.google.common.collect.ImmutableList;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.rbellogger.data.facet.*;
 import de.gematik.rbellogger.util.RbelPathExecutor;
+import de.gematik.test.tiger.LocalProxyRbelMessageListener;
 import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.common.jexl.TigerJexlExecutor;
-import de.gematik.test.tiger.LocalProxyRbelMessageListener;
 import de.gematik.test.tiger.lib.TigerLibraryException;
 import de.gematik.test.tiger.lib.enums.ModeType;
 import de.gematik.test.tiger.lib.json.JsonChecker;
@@ -22,6 +21,7 @@ import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
+import java.util.function.Predicate;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 import java.util.stream.Collectors;
@@ -29,6 +29,7 @@ import javax.xml.transform.Source;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections.iterators.ReverseListIterator;
 import org.apache.commons.lang3.StringUtils;
 import org.assertj.core.api.Assertions;
 import org.awaitility.core.ConditionTimeoutException;
@@ -54,7 +55,7 @@ public class RbelMessageValidator {
         diffOptionMap.put("txtnormalize", DiffBuilder::normalizeWhitespace);
     }
 
-    private static final List<String> emptyPath = ImmutableList.of("", "/");
+    private static final List<String> emptyPath = List.of("", "/");
 
     @Getter
     protected RbelElement currentRequest;
@@ -113,8 +114,10 @@ public class RbelMessageValidator {
                     "No request with path '" + requestParameter.getPath() + "' found in messages");
             } else {
                 throw new AssertionError(
-                    "No request with path '" + requestParameter.getPath() + "' and rbelPath '" + requestParameter.getRbelPath()
-                        + "' matching '" + StringUtils.abbreviate(requestParameter.getValue(), 300) + "' found in messages");
+                    "No request with path '" + requestParameter.getPath() + "' and rbelPath '"
+                        + requestParameter.getRbelPath()
+                        + "' matching '" + StringUtils.abbreviate(requestParameter.getValue(), 300)
+                        + "' found in messages");
             }
         }
         return candidate.get();
@@ -155,8 +158,9 @@ public class RbelMessageValidator {
                     + "Returning " + warnMsg + " message. This may not be deterministic!");
                 printAllPathsOfMessages(candidateMessages);
             }
-            return Optional.of(requestParameter.isFilterPreviousRequest() ? candidateMessages.get(candidateMessages.size() - 1)
-                : candidateMessages.get(0));
+            return Optional.of(
+                requestParameter.isFilterPreviousRequest() ? candidateMessages.get(candidateMessages.size() - 1)
+                    : candidateMessages.get(0));
         }
 
         if (requestParameter.isFilterPreviousRequest()) {
@@ -164,7 +168,8 @@ public class RbelMessageValidator {
         }
 
         for (final RbelElement candidateMessage : candidateMessages) {
-            final List<RbelElement> pathExecutionResult = new RbelPathExecutor(candidateMessage, requestParameter.getRbelPath()).execute();
+            final List<RbelElement> pathExecutionResult = new RbelPathExecutor(candidateMessage,
+                requestParameter.getRbelPath()).execute();
             if (pathExecutionResult.isEmpty()) {
                 continue;
             }
@@ -185,7 +190,8 @@ public class RbelMessageValidator {
                             + StringUtils.abbreviate(requestParameter.getValue(), 300) + "'");
                     }
                 } catch (final Exception ex) {
-                    log.error("Failure while trying to apply regular expression '" + requestParameter.getValue() + "'!", ex);
+                    log.error("Failure while trying to apply regular expression '" + requestParameter.getValue() + "'!",
+                        ex);
                 }
             }
         }
@@ -234,7 +240,8 @@ public class RbelMessageValidator {
         }
     }
 
-    public void assertAttributeOfCurrentResponseMatches(final String rbelPath, final String value, boolean shouldMatch) {
+    public void assertAttributeOfCurrentResponseMatches(final String rbelPath, final String value,
+        boolean shouldMatch) {
         final String text = findElementsInCurrentResponse(rbelPath).stream()
             .map(RbelElement::getRawStringContent)
             .filter(Objects::nonNull)
@@ -319,7 +326,8 @@ public class RbelMessageValidator {
         compareXMLStructure(test, oracle, diffOptions);
     }
 
-    public void compareXMLStructureOfRbelElement(final RbelElement el, final String oracle, final String diffOptionCSV) {
+    public void compareXMLStructureOfRbelElement(final RbelElement el, final String oracle,
+        final String diffOptionCSV) {
         assertThat(el.hasFacet(RbelXmlFacet.class)).withFailMessage("Node " + el.getKey() + " is not XML").isTrue();
         compareXMLStructure(el.getRawStringContent(), oracle, diffOptionCSV);
     }
@@ -371,7 +379,45 @@ public class RbelMessageValidator {
         }
 
         public RbelElement lastResponse() {
-            return currentResponse;
+            return lastMessageMatching(msg ->
+                msg.hasFacet(RbelResponseFacet.class) || msg.hasFacet(RbelHttpResponseFacet.class));
+        }
+
+        public String lastResponseAsString() {
+            return Optional.ofNullable(lastResponse())
+                .map(RbelElement::getRawStringContent)
+                .orElseThrow(NoSuchElementException::new);
+        }
+
+        public RbelElement lastRequest() {
+            return lastMessageMatching(msg ->
+                msg.hasFacet(RbelRequestFacet.class) || msg.hasFacet(RbelHttpRequestFacet.class));
+        }
+
+        private RbelElement lastMessageMatching(Predicate<RbelElement> testMessage) {
+            final ReverseListIterator backwardsIterator = new ReverseListIterator(
+                LocalProxyRbelMessageListener.getValidatableRbelMessages());
+            while (backwardsIterator.hasNext()) {
+                final RbelElement element = (RbelElement) backwardsIterator.next();
+                if (testMessage.test(element)) {
+                    return element;
+                }
+            }
+            throw new NoSuchElementException();
+        }
+
+        public String lastRequestAsString() {
+            return Optional.ofNullable(lastRequest())
+                .map(RbelElement::getRawStringContent)
+                .orElseThrow(NoSuchElementException::new);
+        }
+
+        public String getValueAtLocationAsString(RbelElement element, String rbelPath) {
+            return element.findElement(rbelPath)
+                .flatMap(el -> el.getFacet(RbelValueFacet.class))
+                .map(RbelValueFacet::getValue)
+                .map(Object::toString)
+                .orElseThrow(NoSuchElementException::new);
         }
     }
 }