+++ ReleaseNotes.md
@@ -15,6 +15,7 @@
 * TGR-685: DirectReverseProxy added
 * TGR-664: WebUi: Copies body of response via click on button
 * TGR-525: Rbel moved into Tiger
+* TGR-687: Rbel-Namespace extended (lastRequest/lastMessage/getValueAtLocationAsString)
 
 ## Breaking changes